requirements.txt
Pillow==4.3.0
pytesseract==0.1.7
beautifulsoup4==4.6.0
requests==2.18.4